refactor: изменен скелет проекта
