+++ app/handlers/set_user_name.py
@@ -1,13 +1,13 @@
 from aiogram import F, Router
 from aiogram.filters import StateFilter
 from aiogram.types import Message, CallbackQuery
-from lexicon.lexicon_ru import LEXICON_SET_USER_NAME, LEXICON_MENU_BUTTONS
-from database.database import users_db
-from keyboards.menu_kb import menu_kb
-from keyboards.user_name_setting import yes_no_name_kb
+from app.lexicon.lexicon_ru import LEXICON_SET_USER_NAME, LEXICON_MENU_BUTTONS
+from app.database.database import users_db
+from app.keyboards.menu_kb import menu_kb
+from app.keyboards.user_name_setting import yes_no_name_kb
 
 from aiogram.fsm.context import FSMContext
-from fsm.fsm import FSM_SET_NAME
+from app.fsm.fsm import FSM_SET_NAME
 
 
 router = Router()
