+++ app/keyboards/user_name_setting.py
@@ -1,6 +1,6 @@
 from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 from aiogram.utils.keyboard import InlineKeyboardBuilder
-from lexicon.lexicon_ru import (
+from app.lexicon.lexicon_ru import (
     LEXICON_SET_USER_NAME,
 )
