+++ app/lexicon/lexicon_ru.py
